Before this I've found that 'bk build' is **not** sufficient for triggering the necessary PyXPCOM registration for *newly added* IDL interfaces. I.e. if you do this: - add a new interface in idl - 'bk build' - run Komodo The interface doesn't work in Komodo. That sucks.
The easiest handle for knowing when to force PyXPCOM registration is where we
run 'regxpcom'. Frankly that should be sufficient, but I've found it necessary
to delete the user's compreg.dat file to force *full* re-registration on next
re-start. This is a bit of a burden for making interface *mods*, but necessary
to avoid the WTF when adding a new iface.

r=ericp


git-svn-id: https://svn.openkomodo.com/repos/openkomodo/trunk@770 b87d16e7-d29f-4ee1-b5ec-817d645d335f