find: Warning the user about skipped paths during "replace in files" because we couldn't determine the file type.
The warnings are row items in the find results tree and the tree in the
"Confirm Replacements" dialog.  Eventually the plan is provide a UI there
(double-click on the warning row, or whatever) to add a file association for
"unrecognized languages" so that users can specify that "replace in files" work
in those file types.

Also, fix "replace in files" breakage from a find system checkin yesterday
(openkomodo r959, komodo ide r15222).

r=toddw


git-svn-id: https://svn.openkomodo.com/repos/openkomodo/trunk@960 b87d16e7-d29f-4ee1-b5ec-817d645d335f