Fix: "Find/Replace in Files" don't use Komodo file associations http://bugs.activestate.com/show_bug.cgi?id=75820
Extend textinfo determination to use `env.langinfo_from_filename` for a passed
in "Runtime Environment" class. The Find Service (koIFindService) provide an
appropriate runtime env class that uses the Komodo language register (effectively the file association prefs) to support this.

A lot of the patch is just adding the "env" runtime environment argument to
APIs to pass the koIFindService's implementation through to the findlib2.py's
usage of textinfo.

r=toddw


git-svn-id: https://svn.openkomodo.com/repos/openkomodo/trunk@1187 b87d16e7-d29f-4ee1-b5ec-817d645d335f