Fix: find in files breaks on a broken symlink http://bugs.activestate.com/show_bug.cgi?id=76038
Two separate fixes:
- the change to koFinder.py ensures that any exceptions coming out of the find
backend like this one will be logged in the Find Results tab (rather than the
find in files just dieing silently
- the change to findlib2.py fixes the actual problem with the broken symlinks
here. The exception was due to the textinfo determination failing on
os.stat(path) where "path" was a broken symlink. We trap that and yield a
"SkipPath" event that the Komodo integration can deal with as usual (in this
case it just ignores it).

r=toddw


git-svn-id: https://svn.openkomodo.com/repos/openkomodo/trunk@1249 b87d16e7-d29f-4ee1-b5ec-817d645d335f