fix: New find dialog: Tabstop sequence should go from pattern field to replacement field http://bugs.activestate.com/show_bug.cgi?id=75931
Can't use tabindex because of the mozilla bug 306058, so the "Replace" checkbox
needs to go back down to the lower row (as it was for 4.3.0). However, this
sticks with *disabling* the "Replace with:" textbox (as in 4.3.1) instead of
collapsing it (as in 4.3.0) -- hopefully a satisfactory answer for bug 75737.

r=toddw

(integrated from komodo (trunk) change 17361 by trentm)

git-svn-id: https://svn.openkomodo.com/repos/openkomodo/trunk@1251 b87d16e7-d29f-4ee1-b5ec-817d645d335f