Fix (at least partially): slow typing response (mac os x) http://bugs.activestate.com/show_bug.cgi?id=80683
When Komodo's main toolbox overflowed onto two or more rows on Mac OS X,
Komodo gets pathologically slow. The fix/workaround here is to not
allow this wrapping. Instead, overflow toolbar buttons just get clipped.

Obviously this is a bit of a regression that we'll try to handle (tho
in a different way) in subsequent checkins.

r=trentm (patch by todd)

(integrated from komodo (trunk) change 24858 by trentm)

git-svn-id: https://svn.openkomodo.com/repos/openkomodo/trunk@2578 b87d16e7-d29f-4ee1-b5ec-817d645d335f