Fix for the cursor icon not getting set for the splitter when moving from scintilla (i.e. an editor pane) to the splitter for sidebars or bottom pane.
Shane's guess at the problem is that the "mouseenter" event doesn't get sent to
the element immediatly adjacent to the SciMoz plugin when moving the mouse out
... and that that event is probably what Mozilla is using to change the cursor
icon for the splitter. Adding 1px (transparent, to not look too ugly) around
the plugin fixes that.

This is the first part of the fix/workaround for:
http://bugs.activestate.com/show_bug.cgi?id=80756

r=shanec

(integrated from komodo (trunk) change 26975 by trentm)

git-svn-id: https://svn.openkomodo.com/repos/openkomodo/trunk@3055 b87d16e7-d29f-4ee1-b5ec-817d645d335f