Debugger: (Likely) fix for http://bugs.activestate.com/show_bug.cgi?id=88439 and http://bugs.activestate.com/show_bug.cgi?id=89280 "Komodo hangs when quickly starting/stopping debugging sessions" (88439) "Komodo hangs on Linux after program starts background program and stop/restarting program" (89280)
Most likely, the hang is due to the terminal routine waiting for the process
to finish writing output to the terminal.  While the main process has
ended, the stdio channels are still open due to the child zombie processes,
and Komodo is waiting for output from them.

It seems correct to have Komodo wait at most 5 seconds for any
further output, and then hand the "terminal" (i.e. the command
output window) over to the next process that wants it.

(integrated from komodo (trunk) change 60786 by ericp)

git-svn-id: https://svn.openkomodo.com/repos/openkomodo/trunk@7211 b87d16e7-d29f-4ee1-b5ec-817d645d335f