Editor: Correctly handle indentation after a line containing only leading white-space followed by a line-end comment. The problem is that if we have code like so:
if something: # blah... press shift-return to continue comment here
                        # another shift-return
                        # press regular return here

We need to indent the fourth line relative to the 1st, not the 3rd.  Plus, we need to allow
for blank/empty lines between the current line and the line where the hanging comments
started.  This code is better than fix r72181 for bug 94508 

r=marky
rn=

(integrated from komodo (trunk) change 72247 by ericp)

git-svn-id: https://svn.openkomodo.com/repos/openkomodo/trunk@10380 b87d16e7-d29f-4ee1-b5ec-817d645d335f