find: findbar: make textbox commands (e.g. left, right, home) work again http://bugs.activestate.com/show_bug.cgi?id=96527
The textbox had a controller forwarding to the window, which breaks things
when it has entries for cmd_charPrevious and friends.  Instead, tell the
forwarding controller to prefer the textbox's own controller.

This requires the forwarding controller to deal with loops where it can end
up asking itself for support; check for that case and prevent recursion.

(integrated from komodo (trunk) change 76271 by marky)

git-svn-id: https://svn.openkomodo.com/repos/openkomodo/trunk@11449 b87d16e7-d29f-4ee1-b5ec-817d645d335f